+++ docs/conf.py
@@ -22,7 +22,6 @@ import sys
 # sys.path.insert(0, os.path.abspath('.'))
 sys.path.insert(0, os.path.abspath('..'))
 
-import pypac
 from pypac import __version__
 
 # -- General configuration ------------------------------------------------